lots of updates to B2DstMu_skimCAND_v1.py
This commit contains lots of small updates to B2DstMu_skimCAND_v1.py to get it
working:

- print to stderr when possible.

This makes it easier to see errors in the condor job outputs.

- load all trees and only loop over the events we want.

Previously the code would only loop over the files which contained the events
it was supposed to process. In principle this is more efficient, but it
required quite a bit of index gymnastics to get it right and I was having a
hard time following how it worked. I looked up the source code for
TChain::GetEntry() and it already has all the code to find the correct entry and
(I think?) will avoid loading trees it doesn't need. There is a bit of overhead
each time you call GetEntry() because it has to do this calculation, but since
we are looping over the events in python, I think this overhead is probably
very small, and the code no longer has to figure this out.

- use with open() as f in a few places

I was having jobs fail with a handful of missing input pickle files. I still
don't know exactly what was causing this, but previously the files were never
explicitly closed so this could be the reason.

- use full path to the temporary directory

This allows you to run the script from any directory.

With these changes I was able to successfully run the script on the
BdToDstarMuNu MC files. It took about 2 min for each selection, so not too
long.