speed up trigger scale factors and use iat
This commit speeds up the trigger scale factors calculation by computing
whether events are in between the first and last bincenters before
entering the for loop. Also, switches to using iat[] instead of at[]
since the latter uses indexes which won't work if you apply any cuts and
the index no longer matches the order in the array.